refactor: Merge app-types and types
