+++ packages/neuron-wallet/src/controllers/transactions.ts
@@ -1,4 +1,4 @@
-import { Transaction } from '../app-types/types'
+import { Transaction } from '../types/types'
 import TransactionsService, { PaginationResult, TransactionsByLockHashesParam } from '../services/transactions'
 
 import AddressService from '../services/addresses'
