+++ packages/neuron-wallet/src/services/sync-blocks.ts
@@ -2,13 +2,13 @@ import { getConnection } from 'typeorm'
 import { remote } from 'electron'
 import { Subject, BehaviorSubject } from 'rxjs'
 import Core from '@nervosnetwork/ckb-sdk-core'
-import { Script, OutPoint, Cell, Input, Transaction, Block, BlockHeader } from '../app-types/types'
+import { Script, OutPoint, Cell, Input, Transaction, Block, BlockHeader } from '../types/types'
 import TransactionsService from './transactions'
 import OutputEntity from '../database/chain/entities/output'
 import SyncInfoEntity from '../database/chain/entities/sync-info'
 import NodeService from './node'
 import LockUtils from '../utils/lock-utils'
-import TypeConvert from '../app-types/type-convert'
+import TypeConvert from '../types/type-convert'
 import { NetworkWithID } from './networks'
 import AddressesUsedSubject from '../models/subjects/addresses-used-subject'
