+++ packages/neuron-wallet/src/services/sync/check-and-save/index.ts
@@ -1,4 +1,4 @@
-import { Block } from '../../../app-types/types'
+import { Block } from '../../../types/types'
 import CheckTx from './tx'
 
 export default class CheckAndSave {
