+++ packages/neuron-wallet/src/services/sync/check-and-save/tx.ts
@@ -1,6 +1,6 @@
 import { getConnection } from 'typeorm'
 import { Subject } from 'rxjs'
-import { Transaction, Cell, OutPoint } from '../../../app-types/types'
+import { Transaction, Cell, OutPoint } from '../../../types/types'
 import OutputEntity from '../../../database/chain/entities/output'
 import TransactionsService from '../../transactions'
 import CheckOutput from './output'
