+++ packages/neuron-wallet/src/services/wallets.ts
@@ -10,8 +10,8 @@ import FileService from './file'
 import LockUtils from '../utils/lock-utils'
 import windowManager from '../models/window-manager'
 import { Channel, ResponseCode } from '../utils/const'
-import { Witness, TransactionWithoutHash, Input } from '../app-types/types'
-import ConvertTo from '../app-types/convert-to'
+import { Witness, TransactionWithoutHash, Input } from '../types/types'
+import ConvertTo from '../types/convert-to'
 import Blake2b from '../utils/blake2b'
 import { CurrentWalletNotSet, WalletNotFound, IsRequired, UsedName } from '../exceptions'
 import AddressService from './addresses'
